# contracts/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Resilience API",
    "description": "Deterministic application boundary for the Resilience prototype.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/savings-goals": {
      "get": {
        "tags": [
          "savings-goals"
        ],
        "summary": "List Goals",
        "operationId": "list_goals_api_v1_savings_goals_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavingsGoalListResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "savings-goals"
        ],
        "summary": "Create Goal",
        "operationId": "create_goal_api_v1_savings_goals_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SavingsGoalCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavingsGoalResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/savings-goals/{goal_id}": {
      "patch": {
        "tags": [
          "savings-goals"
        ],
        "summary": "Patch Goal",
        "operationId": "patch_goal_api_v1_savings_goals__goal_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "goal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Goal Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SavingsGoalPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavingsGoalResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "savings-goals"
        ],
        "summary": "Delete Goal",
        "operationId": "delete_goal_api_v1_savings_goals__goal_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "goal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Goal Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/savings-goals/{goal_id}/contributions": {
      "post": {
        "tags": [
          "savings-goals"
        ],
        "summary": "Add Contribution",
        "operationId": "add_contribution_api_v1_savings_goals__goal_id__contributions_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "goal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Goal Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SavingsGoalContributionCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavingsGoalResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/savings-goals/{goal_id}/contributions/{contribution_id}": {
      "delete": {
        "tags": [
          "savings-goals"
        ],
        "summary": "Delete Contribution",
        "operationId": "delete_contribution_api_v1_savings_goals__goal_id__contributions__contribution_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "goal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Goal Id"
            }
          },
          {
            "name": "contribution_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Contribution Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/resilience-jar/summary": {
      "get": {
        "tags": [
          "resilience-jar"
        ],
        "summary": "Get Summary",
        "operationId": "get_summary_api_v1_resilience_jar_summary_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/resilience-jar/plan": {
      "patch": {
        "tags": [
          "resilience-jar"
        ],
        "summary": "Patch Plan",
        "operationId": "patch_plan_api_v1_resilience_jar_plan_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlanPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/resilience-jar/opening-balance": {
      "put": {
        "tags": [
          "resilience-jar"
        ],
        "summary": "Set Opening Balance",
        "operationId": "set_opening_balance_api_v1_resilience_jar_opening_balance_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OpeningBalanceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/resilience-jar/contributions": {
      "post": {
        "tags": [
          "resilience-jar"
        ],
        "summary": "Create Contribution",
        "operationId": "create_contribution_api_v1_resilience_jar_contributions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContributionWrite"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/resilience-jar/withdrawals": {
      "post": {
        "tags": [
          "resilience-jar"
        ],
        "summary": "Create Withdrawal",
        "operationId": "create_withdrawal_api_v1_resilience_jar_withdrawals_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContributionWrite"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/resilience-jar/contributions/{contribution_id}": {
      "patch": {
        "tags": [
          "resilience-jar"
        ],
        "summary": "Update Contribution",
        "operationId": "update_contribution_api_v1_resilience_jar_contributions__contribution_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "contribution_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Contribution Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContributionPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "resilience-jar"
        ],
        "summary": "Delete Contribution",
        "operationId": "delete_contribution_api_v1_resilience_jar_contributions__contribution_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "contribution_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Contribution Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AmountGoalInput": {
        "properties": {
          "mode": {
            "type": "string",
            "const": "amount",
            "title": "Mode"
          },
          "amount_cents": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Amount Cents"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "mode",
          "amount_cents"
        ],
        "title": "AmountGoalInput"
      },
...
      "ContributionPatch": {
        "properties": {
          "amount_cents": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Amount Cents"
          },
          "contribution_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contribution Date"
          },
          "note": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "ContributionPatch"
      },
      "ContributionWrite": {
        "properties": {
          "amount_cents": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Amount Cents"
          },
          "contribution_date": {
            "type": "string",
            "format": "date",
            "title": "Contribution Date"
          },
          "note": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "amount_cents",
          "contribution_date"
        ],
        "title": "ContributionWrite"
      },
      "CoverageGoalInput": {
        "properties": {
          "mode": {
            "type": "string",
            "const": "coverage",
            "title": "Mode"
          },
          "weeks": {
            "type": "integer",
            "maximum": 52.0,
            "minimum": 1.0,
            "title": "Weeks"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "mode",
          "weeks"
        ],
        "title": "CoverageGoalInput"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OpeningBalanceRequest": {
        "properties": {
          "amount_cents": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Amount Cents"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "amount_cents"
        ],
        "title": "OpeningBalanceRequest"
      },
      "PlanPatch": {
        "properties": {
          "recommendation_method": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RecommendationMethod"
              },
              {
                "type": "null"
              }
            ]
          },
          "target_frequency": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TargetFrequency"
              },
              {
                "type": "null"
              }
            ]
          },
          "target_amount_cents": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Amount Cents"
          },
          "weekly_target_cents": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Weekly Target Cents"
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PlanStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "goal": {
            "anyOf": [
              {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/AmountGoalInput"
                  },
                  {
                    "$ref": "#/components/schemas/CoverageGoalInput"
                  }
                ],
                "discriminator": {
                  "propertyName": "mode",
                  "mapping": {
                    "amount": "#/components/schemas/AmountGoalInput",
                    "coverage": "#/components/schemas/CoverageGoalInput"
                  }
                }
              },
              {
                "type": "null"
              }
            ],
            "title": "Goal"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "PlanPatch"
      },
      "PlanStatus": {
        "type": "string",
        "enum": [
          "active",
          "paused"
        ],
        "title": "PlanStatus"
      },
...
      "RecommendationMethod": {
        "type": "string",
        "enum": [
          "conservative_4_week",
          "latest_week"
        ],
        "title": "RecommendationMethod"
      },
...
      "SavingsGoalContributionCreate": {
        "properties": {
          "amountCents": {
            "type": "integer",
            "maximum": 100000000.0,
            "exclusiveMinimum": 0.0,
            "title": "Amountcents"
          },
          "contributedOn": {
            "type": "string",
            "format": "date",
            "title": "Contributedon"
          },
          "note": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "type": "object",
        "required": [
          "amountCents",
          "contributedOn"
        ],
        "title": "SavingsGoalContributionCreate"
      },
      "SavingsGoalContributionResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "amountCents": {
            "type": "integer",
            "title": "Amountcents"
          },
          "contributedOn": {
            "type": "string",
            "format": "date",
            "title": "Contributedon"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          }
        },
        "type": "object",
        "required": [
          "id",
          "amountCents",
          "contributedOn",
          "note",
          "createdAt"
        ],
        "title": "SavingsGoalContributionResponse"
      },
      "SavingsGoalCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "Name"
          },
          "targetCents": {
            "type": "integer",
            "maximum": 100000000.0,
            "exclusiveMinimum": 0.0,
            "title": "Targetcents"
          },
          "targetDate": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Targetdate"
          }
        },
        "type": "object",
        "required": [
          "name",
          "targetCents"
        ],
        "title": "SavingsGoalCreate"
      },
      "SavingsGoalListResponse": {
        "properties": {
          "goals": {
            "items": {
              "$ref": "#/components/schemas/SavingsGoalResponse"
            },
            "type": "array",
            "title": "Goals"
          }
        },
        "type": "object",
        "required": [
          "goals"
        ],
        "title": "SavingsGoalListResponse"
      },
      "SavingsGoalPatch": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 80,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "targetCents": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 100000000.0,
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Targetcents"
          },
          "targetDate": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Targetdate"
          },
          "status": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "active",
                  "completed",
                  "archived"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          }
        },
        "type": "object",
        "title": "SavingsGoalPatch"
      },
      "SavingsGoalResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "targetCents": {
            "type": "integer",
            "title": "Targetcents"
          },
          "targetDate": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Targetdate"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "completed",
              "archived"
            ],
            "title": "Status"
          },
          "savedCents": {
            "type": "integer",
            "title": "Savedcents"
          },
          "remainingCents": {
            "type": "integer",
            "title": "Remainingcents"
          },
          "reached": {
            "type": "boolean",
            "title": "Reached"
          },
          "suggestedWeeklyCents": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Suggestedweeklycents"
          },
          "contributions": {
            "items": {
              "$ref": "#/components/schemas/SavingsGoalContributionResponse"
            },
            "type": "array",
            "title": "Contributions"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "title": "Updatedat"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "targetCents",
          "targetDate",
          "status",
          "savedCents",
          "remainingCents",
          "reached",
          "suggestedWeeklyCents",
          "contributions",
          "createdAt",
          "updatedAt"
        ],
        "title": "SavingsGoalResponse"
      },
...
      "TargetFrequency": {
        "type": "string",
        "enum": [
          "weekly",
          "monthly"
        ],
        "title": "TargetFrequency"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}